% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicw.R
\name{annotateHits}
\alias{annotateHits}
\title{Annotates Hits by Genomic Features}
\usage{
annotateHits(Hits, promoterRange = 1e+05, upstream = NULL,
  downstream = NULL, gene.body = T, promoters = T, utr = T, eqtl = T,
  enhancers = T, loopDist = 0)
}
\description{
Annotates Hits by Genomic Features
}
