% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicw.R
\docType{package}
\name{ComplexID}
\alias{ComplexID}
\alias{ComplexID-package}
\title{ComplexID: Integrated Hit Annotations and Random Walk of Complex Network}
\description{
A package with a self-contained annotation database and protein complex network for finding causitive genes for a phenotype of interest.
}
\section{ComplexID functions}{

runComplexID
}

\author{
Peter Ryabinin
}
