% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicw.R
\name{getEnhancerTissues}
\alias{getEnhancerTissues}
\title{Lists tissues available for enhancer regions}
\usage{
getEnhancerTissues()
}
\value{
A character vector
}
\description{
Lists the tissues that can be used for the enhancerTissues argument in the runComplexID and annotateHits functions
}
\details{
Lists which tissues can be used for annotating enhancer hits.
}
\examples{
getEnhancerTissues()
}
