% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicw.R
\name{geteQTLTissues}
\alias{geteQTLTissues}
\title{Lists tissues available for eQTL sites}
\usage{
geteQTLTissues()
}
\value{
A character vector
}
\description{
Lists the tissues that can be used for the eqtlTissues argument in the runComplexID and annotateHits functions
}
\details{
Lists which tissues can be used for annotating eQTL hits.
}
\examples{
geteQTLTissues()
}
