% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicw.R
\name{getPromoterTissues}
\alias{getPromoterTissues}
\title{Lists tissues available for promoter regions}
\usage{
getPromoterTissues()
}
\value{
A character vector
}
\description{
Lists the tissues that can be used for the promoterTissues argument in the runComplexID and annotateHits functions
}
\details{
Lists which tissues can be used for annotating promoter hits.
}
\examples{
getPromoterTissues()
}
