% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledat.R
\docType{data}
\name{hits.pheno}
\alias{hits.pheno}
\title{Example phenotype similarity matrix for ComplexID}
\format{An object of class \code{matrix} with 8 rows and 2 columns.}
\usage{
hits.pheno
}
\description{
Matrix of similarity values between example phenotypes and phenotpye of interest. \cr
For use with example "hits" data.
}
\keyword{datasets}
