% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledat.R
\docType{data}
\name{hits}
\alias{hits}
\title{Example hits for ComplexID}
\format{An object of class \code{GRanges} of length 160.}
\usage{
hits
}
\description{
160 example hits for using with ComplexID
}
\keyword{datasets}
